♻️ refactor: assets 폴더 분할 및 경로 수정
