+++ src/assets/landing/home_bottom.png
