+++ src/pages/Index.jsx
@@ -2,15 +2,15 @@ import { Link, useNavigate } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
 import Nav from '../components/Nav';
 import Button from '../components/Button';
-import home01 from '../assets/home_01.png';
-import home02 from '../assets/home_02.png';
-import home03 from '../assets/home_03.png';
-import homeBottom from '../assets/home_bottom.png';
-import homeTop from '../assets/home_top.png';
-import icFacebook from '../assets/ic_facebook.svg';
-import icTwitter from '../assets/ic_twitter.svg';
-import icYoutube from '../assets/ic_youtube.svg';
-import icInstagram from '../assets/ic_instagram.svg';
+import home01 from '../assets/landing/home_01.png';
+import home02 from '../assets/landing/home_02.png';
+import home03 from '../assets/landing/home_03.png';
+import homeBottom from '../assets/landing/home_bottom.png';
+import homeTop from '../assets/landing/home_top.png';
+import icFacebook from '../assets/sns/ic_facebook.svg';
+import icTwitter from '../assets/sns/ic_twitter.svg';
+import icYoutube from '../assets/sns/ic_youtube.svg';
+import icInstagram from '../assets/sns/ic_instagram.svg';
 import './Index.css';
 
 function Index() {
